feat: Add Click-to-Click Connector Creation for Image Elements
Implemented a new test case to validate the creation of connectors between image elements using a click-to-click mode. The test ensures that connectors are correctly established when clicking on two images, verifying the state transitions and connector properties throughout the process. This enhancement improves the functionality of the connector system, allowing for more intuitive user interactions with image elements on the board.